resources/conversion/requirements.txt
-f https://download.pytorch.org/whl/torch/
torch==2.4.0+cpu
-f https://download.pytorch.org/whl/torchvision/
torchvision==0.19.0+cpu
-f https://download.pytorch.org/whl/torchaudio/
torchaudio==2.4.0+cpu
torch_xla==2.4.0
--pre
tf-nightly>=2.18.0.dev20240905
ai-edge-litert-nightly
ai-edge-quantizer-nightly
scipy
numpy
tabulate
safetensors
kagglehub
transformers
ai-edge-torch-nightly